NLP/txt_preprocessing/requirements.txt
beautifulsoup4
emoji
nltk
pyspellchecker
